pyproject.toml
[project]
name = "marvin-base"
version = "0.1.0"
description = "Base station nodes for Marvin the droid"
requires-python = ">=3.11"
dependencies = [
  "numpy",
  "pyaudio",
  "pydantic",
  "pydantic-ai",
  "psutil",
  "tornado",
  "pymongo",
  "pillow",
  "cbor2",
  "pocket-tts",
  "lark",
  "torch",
  "torchvision",
  "accelerate",
  "librosa",
  "python-socketio[asyncio_client]",
  "uvicorn",
  "starlette",
  "scipy",
  "pywhispercpp",
  "transformers",
]

[project.optional-dependencies]
dev = [
  "ruff",
  "pyright",
  "pytest",
  "pytest-asyncio",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "standard"
include = ["src"]
